pyproject.toml
[tool.poetry]
name = "Flask-model-validation"
version = "0.1.0"
description = "Extension for Flask-SQLALchemy that include validation for models, a easier way to query and more new features."
authors = ["sr1871 <sergioal18v@gmail.com>"]
readme = "README.md"
repository = "https://github.com/sr1871/flask-model-validation"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python"
]


[tool.poetry.dependencies]
python = "^3.11"
flask-sqlalchemy = "^3.0.3"
python-slugify = "^8.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pyright = "^1.1.308"
pylint = "^2.17.4"
pytest-cov = "^4.0.0"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
